+++ src/features/Conversation/Messages/Assistant/useMarkdown.tsx
@@ -1,72 +1,17 @@
 'use client';
 
-import { type MarkdownProps } from '@lobehub/ui';
 import isEqual from 'fast-deep-equal';
-import { type ReactNode, useMemo, useState } from 'react';
 
-import { HtmlPreviewDrawer } from '@/components/HtmlPreview';
-import { useUserStore } from '@/store/user';
-import { userGeneralSettingsSelectors } from '@/store/user/selectors';
-
-import { markdownElements } from '../../Markdown/plugins';
 import { dataSelectors, messageStateSelectors, useConversationStore } from '../../store';
+import { useChatMarkdown } from '../useChatMarkdown';
 
-const rehypePlugins = markdownElements.map((element) => element.rehypePlugin).filter(Boolean);
-const remarkPlugins = markdownElements.map((element) => element.remarkPlugin).filter(Boolean);
-
-export const useMarkdown = (
-  id: string,
-): { drawer: ReactNode; markdownProps: Partial<MarkdownProps> } => {
+export const useMarkdown = (id: string) => {
   const item = useConversationStore(dataSelectors.getDbMessageById(id), isEqual)!;
-  const { role, search } = item || {};
-  const { transitionMode } = useUserStore(userGeneralSettingsSelectors.config);
-  const generating = useConversationStore(messageStateSelectors.isMessageGenerating(id));
-  const animated = transitionMode === 'fadeIn' && generating;
+  const isGenerating = useConversationStore(messageStateSelectors.isMessageGenerating(id));
 
-  const [drawerContent, setDrawerContent] = useState<string | null>(null);
-
-  const components = useMemo(
-    () =>
-      Object.fromEntries(
-        markdownElements.map((element) => {
-          const Component = element.Component;
-          return [element.tag, (props: any) => <Component {...props} id={id} />];
-        }),
-      ),
-    [id],
-  );
-
-  const markdownProps = useMemo(
-    () =>
-      ({
-        animated,
-        citations: search?.citations,
-        componentProps: {
-          html: {
-            onExpand: (content: string) => setDrawerContent(content),
-          },
-        },
-        components,
-        enableCustomFootnotes: true,
-        enableHtmlPreview: true,
-        enableStream: true,
-        rehypePlugins,
-        remarkPlugins,
-        showFootnotes:
-          search?.citations &&
-          search?.citations.length > 0 &&
-          search?.citations.every((item) => item.title !== item.url),
-      }) satisfies Partial<MarkdownProps>,
-    [animated, components, role, search],
-  );
-
-  const drawer = drawerContent ? (
-    <HtmlPreviewDrawer
-      content={drawerContent}
-      open={!!drawerContent}
-      onClose={() => setDrawerContent(null)}
-    />
-  ) : null;
-
-  return useMemo(() => ({ drawer, markdownProps }), [drawer, markdownProps]);
+  return useChatMarkdown({
+    citations: item?.search?.citations,
+    id,
+    isGenerating,
+  });
 };
