+++ src/features/Conversation/Messages/AssistantGroup/useMarkdown.tsx
@@ -1,48 +1,14 @@
-import { type MarkdownProps } from '@lobehub/ui';
-import { useMemo } from 'react';
-
-import { useUserStore } from '@/store/user';
-import { userGeneralSettingsSelectors } from '@/store/user/selectors';
-
-import { type MarkdownElement } from '../../Markdown/plugins';
-import { markdownElements } from '../../Markdown/plugins';
 import { messageStateSelectors, useConversationStore } from '../../store';
+import { useChatMarkdown } from '../useChatMarkdown';
 
-const rehypePlugins = markdownElements
-  .map((element: MarkdownElement) => element.rehypePlugin)
-  .filter(Boolean);
-const remarkPlugins = markdownElements
-  .map((element: MarkdownElement) => element.remarkPlugin)
-  .filter(Boolean);
-
-export const useMarkdown = (id: string, disableStreaming = false): Partial<MarkdownProps> => {
-  const { transitionMode } = useUserStore(userGeneralSettingsSelectors.config);
-  const generating = useConversationStore(messageStateSelectors.isAssistantGroupItemGenerating(id));
-
-  const enableStream = !disableStreaming;
-  const animated = enableStream && transitionMode === 'fadeIn' && generating;
-
-  const components = useMemo(
-    () =>
-      Object.fromEntries(
-        markdownElements.map((element: MarkdownElement) => {
-          const Component = element.Component;
-
-          return [element.tag, (props: any) => <Component {...props} id={id} />];
-        }),
-      ),
-    [id],
+export const useMarkdown = (id: string, disableStreaming = false) => {
+  const isGenerating = useConversationStore(
+    messageStateSelectors.isAssistantGroupItemGenerating(id),
   );
-  return useMemo(
-    () =>
-      ({
-        animated,
-        components,
-        enableCustomFootnotes: true,
-        enableStream,
-        rehypePlugins,
-        remarkPlugins,
-      }) satisfies Partial<MarkdownProps>,
-    [animated, components, enableStream],
-  );
+
+  return useChatMarkdown({
+    enableStream: !disableStreaming,
+    id,
+    isGenerating,
+  });
 };
